Tratando exceções

+++ src/main/java/com/MarceloAugustoBra/CRUD_clients/services/ClientService.java
@@ -3,6 +3,8 @@ package com.MarceloAugustoBra.CRUD_clients.services;
 import com.MarceloAugustoBra.CRUD_clients.dto.ClientDTO;
 import com.MarceloAugustoBra.CRUD_clients.entities.Client;
 import com.MarceloAugustoBra.CRUD_clients.repositories.ClientRepository;
+import com.MarceloAugustoBra.CRUD_clients.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -17,7 +19,8 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDTO findById(Long id){
-        Client client = repository.findById(id).orElseThrow();
+        Client client = repository.findById(id).orElseThrow(()->
+                new ResourceNotFoundException("Recurso não econtrado!"));
         return new ClientDTO(client);
     }
 
@@ -30,32 +33,39 @@ public class ClientService {
     @Transactional
     public ClientDTO insert (ClientDTO dto){
         Client client = new Client();
-        client.setName(dto.getName());
-        client.setCpf(dto.getCpf());
-        client.setIncome(dto.getIncome());
-        client.setBirthDate(dto.getBirthDate());
-        client.setChildren(dto.getChildren());
-
+        copyDtoToEntity(client, dto);
         client = repository.save(client);
         return new ClientDTO(client);
     }
 
+    @Transactional
     public ClientDTO update (Long id, ClientDTO dto){
-        Client client = repository.getReferenceById(id);
-        client.setName(dto.getName());
-        client.setCpf(dto.getCpf());
-        client.setIncome(dto.getIncome());
-        client.setBirthDate(dto.getBirthDate());
-        client.setChildren(dto.getChildren());
-
-        client = repository.save(client);
-        return new ClientDTO(client);
+       try{
+           Client client = repository.getReferenceById(id);
+           copyDtoToEntity(client, dto);
+           client = repository.save(client);
+           return new ClientDTO(client);
+       }
+       catch (EntityNotFoundException e){
+           throw new ResourceNotFoundException("Recurso não econtrado!");
+       }
 
     }
 
     @Transactional
     public void delete(Long id){
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
         repository.deleteById(id);
+    }
+
+    private void copyDtoToEntity(Client client, ClientDTO dto) {
+        client.setName(dto.getName());
+        client.setCpf(dto.getCpf());
+        client.setIncome(dto.getIncome());
+        client.setBirthDate(dto.getBirthDate());
+        client.setChildren(dto.getChildren());
     }
 
 }
